fix(storybook): Fix popover rendering bug in icon search page
