+++ docs-ui/stories/assets/icons/iconSearch.stories.mdx
@@ -1,0 +1,9 @@
+import SearchPanel from './searchPanel';
+
+<Meta title="Assets/Icons" parameters={{hideToc: true}} />
+
+# Icons
+
+**Pro tip:** in addition to icon name, you can also search by keyword. For example, typing either 'checkmark' or 'success' will return IconCheckmark.
+
+<SearchPanel />
