+++ docs-ui/stories/assets/icons/sample.tsx
