+++ docs-ui/stories/changelog/changelog.stories.mdx
@@ -6,6 +6,28 @@ This is where you can find all of the latest updates to Sentry's design and comp
 
 ---
 
+<p className="small">Jan 12, 2022</p>
+
+## Icon search
+
+Quickly find icons in the new search tool, see available props, and get instantly copiable JSX snippets.
+
+<DocsLinks
+  links={[
+    {
+      img: {
+        src: require('docs-ui/images/icons-link-preview.svg'),
+        alt: 'Large colorful icons',
+      },
+      title: 'Icons',
+      kind: 'Assets/Icons',
+      story: 'Page',
+    },
+  ]}
+/>
+
+---
+
 <p className="small">Nov 9, 2021</p>
 
 ## New Colors page
