--- docs-ui/stories/core/icons/iconSearch.stories.mdx
@@ -1,9 +1,0 @@
-import SearchPanel from './searchPanel';
-
-<Meta title="Core/Icons/Search" parameters={{hideToc: true}} />
-
-# Search Icons
-
-**Pro tip:** you can search by keyword in addition to icon name. For example, typing either 'checkmark' or 'success' will return IconCheckmark.
-
-<SearchPanel />
