--- docs-ui/stories/core/icons/iconUsage.stories.mdx
@@ -1,7 +1,0 @@
-import SearchPanel from './searchPanel';
-
-<Meta title="Core/Icons/Usage" />
-
-# Icons
-
-We've built Sentry's type system around Rubik - a playful open-source typeface. For code and code-like elements, we use Roboto Mono.
