--- docs-ui/stories/core/icons/icons.stories.mdx
@@ -1,9 +1,0 @@
-import SearchPanel from './searchPanel';
-
-<Meta title="Core/Icons" />
-
-# Icons
-
-We've built Sentry's type system around Rubik - a playful open-source typeface. For code and code-like elements, we use Roboto Mono.
-
-<SearchPanel />
