+++ docs-ui/stories/gettingStarted/gettingStarted.stories.mdx
@@ -97,5 +97,15 @@ Start reading now with these foundational topics.
       kind: 'Core/Typography',
       story: 'Page',
     },
+    {
+      img: {
+        src: require('docs-ui/images/icons-link-preview.svg'),
+        alt: 'Large colorful icons',
+      },
+      title: 'Icons',
+      desc: 'Search for icons and get copiable code snippets',
+      kind: 'Assets/Icons',
+      story: 'Page',
+    },
   ]}
 />
