+++ package.json
@@ -122,7 +122,7 @@
     "react-hotkeys-hook": "^3.4.3",
     "react-lazyload": "^2.3.0",
     "react-mentions": "4.3.0",
-    "react-popper": "1",
+    "react-popper": "^1.3.11",
     "react-router": "3.2.0",
     "react-select": "^3.0.8",
     "react-sparklines": "1.7.0",
